type change

+++ GW2EIEvtcParser/Extensions/ExtensionCombatData/ExtensionCombatEvents/HealingStats/EXTAbstractHealingEvent.cs
@@ -6,7 +6,7 @@ namespace GW2EIEvtcParser.Extensions
 {
     public abstract class EXTAbstractHealingEvent : AbstractDamageEvent
     {
-        public long HealingDone { get; protected set; }
+        public int HealingDone { get; protected set; }
 
         internal EXTAbstractHealingEvent(CombatItem evtcItem, AgentData agentData, SkillData skillData) : base(evtcItem, agentData, skillData)
         {
